use a class
